move from local/docker to local/podman
